# swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Movie Catalog API",
      "version": "1.0.0",
      "description": "API for managing a movie catalog"
    },
    "servers": [
      {
        "url": "http://localhost:3000",
        "description": "Local development server"
      }
    ],
    "components": {
      "securitySchemes": {
        "bearerAuth": {
          "type": "http",
          "scheme": "bearer",
          "bearerFormat": "JWT"
        }
      },
      "schemas": {
        "User": {
          "type": "object",
          "properties": {
            "email": {
              "type": "string",
              "format": "email"
            },
            "password": {
              "type": "string",
              "format": "password"
            }
          }
        },
        "Movie": {
          "type": "object",
          "properties": {
            "title": {
              "type": "string"
            },
            "description": {
              "type": "string"
            },
            "releaseDate": {
              "type": "string",
              "format": "date"
            },
            "userId": {
              "type": "integer"
            }
          }
        }
      }
    },
    "paths": {
      "/auth/register": {
        "post": {
          "summary": "Register a new user",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "User registered"
            },
            "400": {
              "description": "Invalid input"
            }
          }
        }
      },
      "/auth/login": {
        "post": {
          "summary": "Login a user",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "User logged in"
            },
            "401": {
              "description": "Invalid credentials"
            }
          }
        }
      },
      "/movies": {
        "post": {
          "summary": "Add a new movie",
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Movie"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Movie added"
            },
            "400": {
              "description": "Invalid input"
            },
            "401": {
              "description": "Unauthorized"
            }
          }
        },
        "get": {
          "summary": "Get all movies",
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "responses": {
            "200": {
              "description": "List of movies"
            },
            "401": {
              "description": "Unauthorized"
            }
          }
        }
      },
      "/movies/{id}": {
        "get": {
          "summary": "Get a movie by ID",
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Movie details"
            },
            "401": {
              "description": "Unauthorized"
            },
            "404": {
              "description": "Movie not found"
            }
          }
        },
        "put": {
          "summary": "Update a movie by ID",
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Movie"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Movie updated"
            },
            "400": {
              "description": "Invalid input"
            },
            "401": {
              "description": "Unauthorized"
            },
            "404": {
              "description": "Movie not found"
            }
          }
        },
        "delete": {
          "summary": "Delete a movie by ID",
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Movie deleted"
            },
            "401": {
              "description": "Unauthorized"
            },
            "404": {
              "description": "Movie not found"
            }
          }
        }
      }
    }
  }
  